[core] new function as move-only & optimized
Through some experimentation it was determined that the most efficient type-erasure method for fast delegates is a combination of in-situ enframed side-band control data and fused operations to reduce the size of that control data. This implementation fuses the move and destruction operation into a single function, while leaving invocation as its own function to avoid UB and wasting stack space for possible function arguments. The enframed data is stored as a header to the functor and is always at constant offsets that avoid straddling cache lines. This implementation makes only one indirection for any given dispatch.